refactor: consolidate duplicated _migrated_db_template fixture
Move the shared _migrated_db_template fixture to tests/conftest.py using
run_migrations() directly, replacing the integration-test variant that
booted a full DatabaseService (with a throwaway asyncio event loop) just
to produce a migrated schema file. Both premigrated_db_path (integration)
and telemetry_db (unit) now consume the single shared fixture.

Closes #1209
